Move IpSpaceSpecializer interface to common
Also found a few Batfish datamodel visitors that seem like they should be there.
