+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/visitors/IpSpaceMayIntersectWildcard.java
