+++ projects/batfish/src/main/java/org/batfish/z3/IpSpaceIpSpaceSpecializer.java
@@ -5,6 +5,8 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import javax.annotation.Nonnull;
+import org.batfish.common.ipspace.IpSpaceSimplifier;
+import org.batfish.common.ipspace.IpSpaceSpecializer;
 import org.batfish.datamodel.EmptyIpSpace;
 import org.batfish.datamodel.Ip;
 import org.batfish.datamodel.IpSpace;
